fix(Next): Separate api from library source
