--- __mocks__/react-mapbox-gl.js
@@ -1,1 +1,0 @@
-module.exports = () => jest.fn(() => {});
